Updated build.sh

+++ build.sh
@@ -1,5 +1,11 @@
-#!/bin/sh
+#!/bin/bash
+
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
+
 php build.php
-mv insightly-cli.phar /usr/local/bin/isc
-chmod +x /usr/local/bin/isc
+chmod +x ./insightly-cli.phar
+mv insightly-cli.phar /usr/bin/isc
 rm insightly-cli.phar.gz
